+++ NeteaseDX9HW/ArcMath.h
@@ -215,7 +215,7 @@ inline float3 getSphereCenter(float x1, float y1, float z1,
 	float x3, float y3, float z3,
 	float x4, float y4, float z4)//空间四点确定球心坐标(克莱姆法则)
 {
-	double a11, a12, a13, a21, a22, a23, a31, a32, a33, b1, b2, b3, d, d1, d2, d3;
+	float a11, a12, a13, a21, a22, a23, a31, a32, a33, b1, b2, b3, d, d1, d2, d3;
 	a11 = 2 * (x2 - x1); a12 = 2 * (y2 - y1); a13 = 2 * (z2 - z1);
 	a21 = 2 * (x3 - x2); a22 = 2 * (y3 - y2); a23 = 2 * (z3 - z2);
 	a31 = 2 * (x4 - x3); a32 = 2 * (y4 - y3); a33 = 2 * (z4 - z3);
